+++ packages/sv/CHANGELOG.md
@@ -1,5 +1,20 @@
 # sv
 
+## 0.16.6
+### Patch Changes
+
+
+- fix: format all created project files when using the prettier add-on ([#1192](https://github.com/sveltejs/cli/pull/1192))
+
+
+- feat: smart(er) package manager selection ([#1190](https://github.com/sveltejs/cli/pull/1190))
+
+
+- feat(cli): `addOption` is now available in the setup phase to dynamically add options to your add-on ([#1042](https://github.com/sveltejs/cli/pull/1042))
+
+- Updated dependencies [[`580c2f7`](https://github.com/sveltejs/cli/commit/580c2f7ae34d733393c4e5c185feffe0f1fe9f8e)]:
+  - @sveltejs/sv-utils@0.3.2
+
 ## 0.16.5
 ### Patch Changes